[feature] 관리자 전시회 등록 완료
