+++ src/components/regexh/RegExhFormFrame.tsx
@@ -37,6 +37,7 @@ import {changeImageSize} from '~/utils/resizeImage';
 import LoadingModal from '~/components/common/modal/LoadingModal';
 import {changeDateTimeFormat, changeDotToHyphen} from '~/utils/date';
 import ExhSelectPeriod from './ExhSelectPeriodModal';
+import {UpdateRegExhByAdminType} from '~/api/regexh';
 
 type RegExhDataSetType = {
   regExhName: string;
@@ -58,7 +59,6 @@ type RegExhDataSetType = {
   regPosterUri: string | undefined;
   setRegPosterUri: (text: string | undefined) => void;
   regComment?: string | undefined; // 관리자 업데이트일 경우에 해당
-  // setRegComment?: (text: string | undefined) => void;
 };
 
 type CreateApiType = {
@@ -67,17 +67,18 @@ type CreateApiType = {
 };
 
 type UpdateByAdminApiType = {
-  updateByAdminApi: (formData: FormData | null) => void;
+  regExhId: number;
+  updateByAdminApi: (updateData: UpdateRegExhByAdminType) => void;
 };
 
 type RequestApiType = {
   isLoading: boolean;
   // 사용자 추가 api
   createRequest?: CreateApiType;
-  // 사용자 업데이트 api
-  // 여기에 추가해주세용
   // 관리자 업데이트 api
   updateByAdminRequest?: UpdateByAdminApiType;
+  // 사용자 업데이트 api
+  //  - 여기에 추가해주세용
 };
 
 interface RegExhFormFrameProps {
@@ -271,7 +272,10 @@ const RegExhFormFrame: React.FC<RegExhFormFrameProps> = ({
       if (regExhData.regComment) {
         formData.append('regComment', regExhData.regComment);
       }
-      requestData.updateByAdminRequest?.updateByAdminApi(formData);
+      requestData.updateByAdminRequest?.updateByAdminApi({
+        regExhId: requestData.updateByAdminRequest?.regExhId,
+        formData,
+      });
     }
   };
 
@@ -354,7 +358,7 @@ const RegExhFormFrame: React.FC<RegExhFormFrameProps> = ({
                 onChangeText={onChangeFee}
                 value={regExhData.regFee}
               />
-              <SectionName>원</SectionName>
+              <SectionName color={'default'}>원</SectionName>
             </FeeWrapper>
           </RowSectionWrapper>
           {/* 전시회 홈페이지 링크 */}
@@ -467,15 +471,18 @@ interface SectionColorProps {
 
 const SectionName = styled.Text<SectionColorProps>`
   font-size: ${AREA_FONT_SIZE}px;
-  color: ${DEFAULT_TEXT};
   font-family: ${FONT_NAME};
   color: ${(props: SectionColorProps) =>
-    props.color === 'grey' ? `${LIGHT_GREY}` : `${DEFAULT_TEXT}`};
+    props.color === 'grey'
+      ? `${LIGHT_GREY}`
+      : props.color === 'default'
+        ? `${DEFAULT_TEXT}`
+        : `${MIDDLE_GREY}`};
 `;
 
 const DateText = styled.Text`
   font-size: ${rf(15.5)}px;
-  color: ${MIDDLE_GREY};
+  color: ${DEFAULT_TEXT};
   font-family: ${FONT_NAME};
 `;
