+++ src/screens/setting/ConfirmRegExhScreen.tsx
@@ -1,25 +1,13 @@
-import React, {useCallback, useEffect, useState} from 'react';
-import styled from 'styled-components/native';
-import {RouteProp, useNavigation} from '@react-navigation/native';
-import {RootStackNavigationProp} from '~/App';
+import React, {useEffect, useState} from 'react';
+import {RouteProp} from '@react-navigation/native';
 import {showToast} from '~/components/common/modal/toastConfig';
-import {useUpdateRegExhByAdmin} from '~/api/queries/regexh';
-import RegExhFormFrame from '~/components/regexh/RegExhFormFrame';
-import {
-  AREA_FONT_SIZE,
-  BUTTON_RADIUS,
-  FONT_NAME,
-  ITEM_BORDER_WIDTH,
-} from '~/components/common/style';
-import {DEFAULT_TEXT, LIGHT_GREY} from '~/components/common/colors';
-import {
-  responseFont as rf,
-  heightSizePercentage as hp,
-  widthSizePercentage as wp,
-} from '~/components/common/ResponsiveSize';
+import {usefetchRegExhDetail} from '~/api/queries/regexh';
+import LoadingModal from '~/components/common/modal/LoadingModal';
+import ConfirmRegExh from './ConfirmRegExh';
+import ShowConfirmedRegExh from './ShowConfirmedRegExh';
 
 type RootStackParamList = {
-  ConfirmRegExhScreen: {rExhId: number};
+  ConfirmRegExhScreen: {regExhId: number};
 };
 
 type ConfirmRegExhScreenProp = RouteProp<
@@ -32,137 +20,51 @@ interface Props {
 }
 
 const ConfirmRegExhScreen: React.FC<Props> = ({route}) => {
-  const navigation = useNavigation<RootStackNavigationProp>();
+  const {regExhId} = route.params;
   // fetch api
-
-  const [regExhName, setRegExhName] = useState<string>('');
-  const [regGallery, setRegGallery] = useState<string>('');
-  const [regStartDate, setRegStartDate] = useState<string>('');
-  const [regEndDate, setRegEndDate] = useState<string>('');
-  const [regPainter, setRegPainter] = useState<string>('');
-  const [regFee, setRegFee] = useState<string>('');
-  const [regUrl, setRegUrl] = useState<string | undefined>(undefined);
-  const [regIntro, setRegIntro] = useState<string | undefined>(undefined);
-  const [regPosterUri, setRegPosterUri] = useState<string | undefined>(
-    undefined,
+  const {data, isLoading, isError, isSuccess, refetch} = usefetchRegExhDetail(
+    regExhId,
+    true,
   );
-  const [regComment, setRegComment] = useState<string | undefined>(undefined);
+  const [regExhInfo, setRegExhInfo] = useState<any | null>(null);
 
-  const [regExhFormdata, setRegExhFormdata] = useState<FormData | null>(null);
+  const [isLoadingOpen, setIsLoadingOpen] = useState<boolean>(false);
 
-  // update by admin api
-  const {
-    mutate: updateRegExh,
-    isLoading: isLoading,
-    isError: isError,
-    isSuccess: isSuccess,
-  } = useUpdateRegExhByAdmin();
+  useEffect(() => {
+    if (isSuccess) {
+      setRegExhInfo(data);
+    }
+  }, [data, isSuccess]);
 
   useEffect(() => {
     if (isError) {
-      showToast('전시회 등록 작성을 실패했습니다.');
+      showToast('전시회 등록 정보 조회를 실패했습니다.');
     }
-    if (isSuccess) {
-      // 설정 페이지의 등록한 전시회 페이지로 이동
+    if (isLoading) {
+      setIsLoadingOpen(true);
+    } else {
+      setIsLoadingOpen(false);
     }
-  }, [isError, isSuccess]);
-
-  const onChangeComment = useCallback((text: string) => {
-    setRegComment(text);
-  }, []);
+  }, [isError, isLoading]);
 
   return (
-    <RegExhFormFrame
-      formState={'updateByAdmin'}
-      regExhData={{
-        regExhName,
-        setRegExhName,
-        regGallery,
-        setRegGallery,
-        regStartDate,
-        setRegStartDate,
-        regEndDate,
-        setRegEndDate,
-        regPainter,
-        setRegPainter,
-        regFee,
-        setRegFee,
-        regUrl,
-        setRegUrl,
-        regIntro,
-        setRegIntro,
-        regPosterUri,
-        setRegPosterUri,
-        regComment,
-      }}
-      requestData={{
-        isLoading: isLoading,
-      }}>
-      <ColSectionWrapper>
-        <SectionView>
-          <SectionName>코멘트</SectionName>
-          <SectionName color={'grey'}>(선택)</SectionName>
-        </SectionView>
-        <RowSectionWrapper>
-          <WriteInfo
-            textAlignVertical={'top'}
-            multiline={true}
-            placeholderTextColor={LIGHT_GREY}
-            placeholder={'코멘트 작성'}
-            value={regComment}
-            onChangeText={onChangeComment}
+    <>
+      {regExhInfo &&
+        (!regExhInfo.regState ? (
+          // regState == false면 등록 화면 보여주기
+          <ConfirmRegExh regExhId={regExhId} regExhInfo={regExhInfo} />
+        ) : (
+          // regState == true면 미리보기 화면 보여주기
+          <ShowConfirmedRegExh
+            regExhId={regExhId}
+            regExhInfo={regExhInfo}
+            refetch={refetch}
           />
-        </RowSectionWrapper>
-      </ColSectionWrapper>
-    </RegExhFormFrame>
+        ))}
+
+      {isLoadingOpen && <LoadingModal message={'정보 불러오는 중 :)'} />}
+    </>
   );
 };
 
 export default ConfirmRegExhScreen;
-
-/** style */
-const RowSectionWrapper = styled.View`
-  flex-direction: row;
-  border-width: ${ITEM_BORDER_WIDTH}px;
-  border-color: ${LIGHT_GREY};
-  border-radius: ${BUTTON_RADIUS}px;
-  align-items: center;
-  justify-content: space-between;
-  padding-left: ${wp(2.8)}px;
-  padding-right: ${wp(2.8)}px;
-`;
-
-const ColSectionWrapper = styled.View`
-  flex-direction: column;
-  border-width: ${ITEM_BORDER_WIDTH}px;
-  border-color: ${LIGHT_GREY};
-  border-radius: ${BUTTON_RADIUS}px;
-  width: 100%;
-  padding: ${wp(2.9)}px;
-  gap: ${hp(1.6)}px;
-`;
-
-const SectionView = styled.View`
-  flex-direction: row;
-  gap: ${hp(0.3)}px;
-`;
-
-interface SectionProps {
-  color: string;
-}
-
-const SectionName = styled.Text<SectionProps>`
-  font-size: ${AREA_FONT_SIZE}px;
-  color: ${DEFAULT_TEXT};
-  font-family: ${FONT_NAME};
-  color: ${(props: SectionProps) =>
-    props.color === 'grey' ? `${LIGHT_GREY}` : `${DEFAULT_TEXT}`};
-`;
-
-const WriteInfo = styled.TextInput`
-  font-size: ${rf(16)}px;
-  color: ${DEFAULT_TEXT};
-  font-family: ${FONT_NAME};
-  padding-right: 0%;
-  min-height: ${hp(25)}px;
-`;
